pybridge/requirements.txt
anthropic>=0.40.0
openai>=1.50.0
requests>=2.31.0
Pillow>=10.0.0
python-telegram-bot>=20.0
psutil>=5.9.0
playwright>=1.40.0
pip-audit>=2.6.0
qrcode[pil]>=7.4
